- Fuji builds better on Linux. - Added MonoDevelop project.
git-svn-id: https://fuji.googlecode.com/svn/trunk@663 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Include/MFPoolHeap.h
@@ -7,7 +7,7 @@ public:
 	void Init(int num, int size, void *pMem = NULL, int memsize = 0);
 	void Destroy();
 
-	uint32 Size() const;
+	size_t Size() const;
 	void *Alloc();
 	void Delete(void *pItem);
 	void DeleteAll();
